+++ src/components/comment-edit-form.jsx
@@ -3,6 +3,7 @@ import { useMemo, useCallback, useState, useRef, useEffect, useContext } from 'r
 import cn from 'classnames';
 
 import { faPaperclip } from '@fortawesome/free-solid-svg-icons';
+import { useSelector } from 'react-redux';
 import { initialAsyncState } from '../redux/async-helpers';
 import { doneEditingAndDeleteDraft, getDraft } from '../services/drafts';
 import { Throbber } from './throbber';
@@ -14,6 +15,7 @@ import { SmartTextarea } from './smart-textarea';
 import { useUploader } from './uploader/uploader';
 import { useFileChooser } from './uploader/file-chooser';
 import { UploadProgress } from './uploader/progress';
+import { PreventPageLeaving } from './prevent-page-leaving';
 
 export function CommentEditForm({
   initialText = '',
@@ -26,6 +28,7 @@ export function CommentEditForm({
   submitStatus = initialAsyncState,
   draftKey,
 }) {
+  const frontendPreferences = useSelector((state) => state.user.frontendPreferences);
   const { setInput } = useContext(PostContext);
   const input = useRef(null);
   const [text, setText] = useState(() => getDraft(draftKey)?.text ?? initialText);
@@ -38,13 +41,16 @@ export function CommentEditForm({
 
   const doCancel = useCallback(
     (e) => {
-      if (text !== initialText && !confirm('Discard changes?')) {
+      if (text.trim() !== initialText.trim() && !confirm('Discard changes?')) {
         return;
+      }
+      if (isPersistent) {
+        setText(initialText);
       }
       onCancel(e);
       doneEditingAndDeleteDraft(draftKey);
     },
-    [draftKey, initialText, onCancel, text],
+    [draftKey, initialText, isPersistent, onCancel, text],
   );
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -96,6 +102,9 @@ export function CommentEditForm({
 
   return (
     <div className="comment-body" role="form">
+      <PreventPageLeaving
+        prevent={!frontendPreferences.saveDrafts && (canSubmit || submitStatus.loading)}
+      />
       <div>
         <SmartTextarea
           ref={input}
@@ -135,16 +144,14 @@ export function CommentEditForm({
           Comment
         </button>
 
-        {!isPersistent && (
-          <ButtonLink
-            className="comment-cancel"
-            onClick={doCancel}
-            aria-disabled={submitStatus.loading}
-            aria-label={submitStatus.loading ? 'Cancel disabled (submitting)' : null}
-          >
-            Cancel
-          </ButtonLink>
-        )}
+        <ButtonLink
+          className="comment-cancel"
+          onClick={doCancel}
+          aria-disabled={submitStatus.loading}
+          aria-label={submitStatus.loading ? 'Cancel disabled (submitting)' : null}
+        >
+          {isPersistent ? 'Clear' : 'Cancel'}
+        </ButtonLink>
 
         <SubmitModeHint input={input} />